Make Midscene Intel build portable

+++ Tinycast/DesignSystem/Scrolling/OverflowFade.swift
@@ -44,22 +44,22 @@ struct OverflowFadeMask: ViewModifier {
         }
         let extent = min(band / height, 0.5)
         return [
-            .init(color: .black.opacity(1 - topStrength), location: 0),
+            .init(color: .black.opacity(Double(1 - topStrength)), location: 0),
             .init(
-                color: .black.opacity(1 - topStrength * 0.75),
+                color: .black.opacity(Double(1 - topStrength * 0.75)),
                 location: extent * 0.35),
             .init(
-                color: .black.opacity(1 - topStrength * 0.25),
+                color: .black.opacity(Double(1 - topStrength * 0.25)),
                 location: extent * 0.7),
             .init(color: .black, location: extent),
             .init(color: .black, location: 1 - extent),
             .init(
-                color: .black.opacity(1 - bottomStrength * 0.25),
+                color: .black.opacity(Double(1 - bottomStrength * 0.25)),
                 location: 1 - extent * 0.7),
             .init(
-                color: .black.opacity(1 - bottomStrength * 0.75),
+                color: .black.opacity(Double(1 - bottomStrength * 0.75)),
                 location: 1 - extent * 0.35),
-            .init(color: .black.opacity(1 - bottomStrength), location: 1)
+            .init(color: .black.opacity(Double(1 - bottomStrength)), location: 1)
         ]
     }
 
@@ -70,7 +70,7 @@ struct OverflowFadeMask: ViewModifier {
         return [
             .init(color: .black, location: 0),
             .init(color: .black, location: 1 - band / height),
-            .init(color: .black.opacity(1 - strength), location: 1)
+            .init(color: .black.opacity(Double(1 - strength)), location: 1)
         ]
     }
 }
